+++ imports/ui/RouterComponents/activities/NewActivityContainer.js
@@ -1,6 +1,7 @@
 import { withTracker } from 'meteor/react-meteor-data';
 import NewActivity from './NewActivity.jsx';
 import { parseActsWithResources } from '../../functions';
+import Resources from '../../../api/resources/resource.js';
 
 export default NewActivityContainer = withTracker((props) => {
   const resourcesSub = Meteor.subscribe('resources');
