+++ lib/collections.js
@@ -7,7 +7,7 @@ Activities = new Mongo.Collection('activities');
 // Works = new Mongo.Collection('works');
 // Pages = new Mongo.Collection('pages');
 
-Resources = new Mongo.Collection('resources');
+// Resources = new Mongo.Collection('resources');
 // Chats = new Mongo.Collection('chats');
 // Documents = new Mongo.Collection('documents');
 // Categories = new Mongo.Collection('categories');
